+++ package.json
@@ -16,7 +16,7 @@
     "tape": "^3.0.3"
   },
   "scripts": {
-    "test": "tape test.js"
+    "test": "tape ./test/test.js"
   },
   "repository": {
     "type": "git",
